% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\encoding{UTF-8}
\name{agregation}
\alias{agregation}
\title{Agregation}
\usage{
agregation(
  donnees,
  unite,
  age,
  age_cat_mins = NULL,
  age_filtre_max = Inf,
  sexe = NULL,
  autres_vars = NULL,
  type_num,
  type_denom,
  num_filtre_expression = NULL,
  numerateur_agr_col = NULL,
  denominateur_agr_col = NULL,
  denom_externe_type_unite = NULL,
  denom_externe_annee = NULL,
  denom_externe_geo = NULL,
  denom_externe_regroupement_unite = NULL
)
}
\arguments{
\item{donnees}{Un dataframe contenant les données à agréger. Les valeurs NA ne sont pas permises.}

\item{age}{Un champ dans le dataframe qui indique l'âge du patient en format continu à regrouper par `age_cat_mins`.}

\item{age_cat_mins}{Un vecteur contenant les âges minimaux de chaque groupe d'âge pour le regroupement de `age` et/ou de la variable d'âge du tableau externe de dénominateur.}

\item{age_filtre_max}{Une valeur numérique indiquant le filtre d'âge maximum exclusif, Inf par défaut.}

\item{sexe}{Un champ dans le dataframe qui indique le sexe de chaque individu. Les valeurs de cette variable doivent être "M" et "F". Facultatif.}

\item{autres_vars}{Un vecteur optionnel d'autres variable(s) pour la stratification. Facultatif.}

\item{type_num}{Une chaîne de caractères indiquant le type de numérateur. Il peut être 'filtré' si les cas sont identifiés par une exression logique (`num_filtre_expression`), 'total_interne' si la somme de rangées représente les cas, ou 'colonne agrégée' si un champ de `donnees`indique les numérateurs agrégées.}

\item{type_denom}{Une chaîne de caractères indiquant le type de dénominateur. Il peut être 'total interne' si la somme de rangées représente les dénominateurs, 'externe' si des dénominateurs externes sont utilisés, ou 'colonne agrégée' si un champ de `donnees`indique les dénominateurs agrégées.}

\item{num_filtre_expression}{Une chaîne de caractères représentant une expression conditionnelle utilisée pour filtrer les données. Cette expression est utilisée lorsque `type_num` est 'filtré'.}

\item{numerateur_agr_col}{Une colonne facultative dans le dataframe `donnees` qui contient des numérateurs agrégés. Ceci n'est utilisé que lorsque `type_num` est 'colonne agrégée'.}

\item{denominateur_agr_col}{Une colonne facultative dans le dataframe `donnees` qui contient des dénominateurs agrégés. Ceci n'est utilisé que lorsque `type_denom` est 'colonne agrégée'.}

\item{denom_externe_type_unite}{Une chaîne de caractères facultative indiquant le type d'unité pour le dénominateur externe. Il peut être 'Régional' ou 'Annuel'.}

\item{denom_externe_annee}{Si `denom_externe_type_unite` = 'Régional', un valeur indiquant l'année pour le dénominateur externe.}

\item{denom_externe_geo}{Si `denom_externe_type_unite` = 'Régional', une chaîne de caractères facultative indiquant le niveau géographique pour le dénominateur externe. Il peut être 'RSS', 'RLS', ou 'RTS'.}

\item{denom_externe_regroupement_unite}{Une liste facultative contenant des correspondances de nouveaux noms de groupe à des noms d'unité originaux pour le dénominateur externe. Ceci est utilisé pour effectuer une correspondance entre les regroupement d'unités de `donnees` et du dataframe pop_ref.}
}
\value{
Un dataframe incluant une colonne pour les unités d'analyse (par exemple les régions ou les années), les variables agrégées (age, sexe, autres), le numérateur et le dénominateur. Le regroupement de la variable d'âge sort toujours avec le nom AGE_CAT
}
\description{
`agregation` est une fonction qui permet de transformer des données individuelles ou partiellement agrégées en un format adapté pour les fonctions de standardisation\cr
}
\examples{
\dontrun{

data_n<-agregation(donnees = donnees_sim \%>\% filter(annee=="2022"),
          unite = "rss",
          age = "age",
          age_cat_mins = c(0,50, 60,70),
          sexe = "sexe",
          type_num = "filtré",
          num_filtre_expression = "deces == 'Oui'",
          type_denom = 'total interne')
          
glimpse(data_n) 
}

}
\keyword{internal}
