% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_sim.R
\docType{data}
\name{donnees_sim}
\alias{donnees_sim}
\title{Données simulées de patients COVID-19 au Québec}
\format{
Un object de classe \code{"tibble"};
}
\usage{
data(donnees_sim)
}
\description{
Données simulées basées sur les distribution d'âge, de sexe, et de décès par 
région RSS d'un échantillon de 10% des patients hospitalisés pour la COVID-19
au Québec de 2020 à 2022.
}
\examples{
data(donnees_sim)
table(donnees_sim$deces)
}
\keyword{datasets}
