% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.agregation.R
\name{plot.agregation}
\alias{plot.agregation}
\title{Méthode Plot pour un objet de classe "agregation"}
\usage{
\method{plot}{agregation}(x)
}
\arguments{
\item{x}{un objet de la classe "agregation"}
}
\description{
Méthode Plot pour un objet de classe "agregation"
}
