% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.standardisation_list.R
\name{plot.standardisation_list}
\alias{plot.standardisation_list}
\title{Méthode Plot pour un objet de classe "standardisation_list"}
\usage{
\method{plot}{standardisation_list}(x, Annuel = F)
}
\arguments{
\item{x}{un objet de la classe "standardisation_list"}

\item{Annuel}{un argument, FALSE par défaut,  pour définir si les unités standardisés sont des années}
}
\description{
Méthode Plot pour un objet de classe "standardisation_list"
}
