% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.agregation.R
\name{print.agregation}
\alias{print.agregation}
\title{Print method for "agregation" class}
\usage{
\method{print}{agregation}(x)
}
\arguments{
\item{x}{an object of class "agregation"}
}
\description{
Print method for "agregation" class
}
