% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.standardisation_list.R
\name{print.standardisation_list}
\alias{print.standardisation_list}
\title{Print method for "standardisation_list" class}
\usage{
\method{print}{standardisation_list}(x)
}
\arguments{
\item{x}{an object of class "standardisation_list"}
}
\description{
Print method for "standardisation_list" class
}
