% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisation_externe.R
\encoding{UTF-8}
\name{standardisation_externe}
\alias{standardisation_externe}
\title{Standardisation externe}
\usage{
standardisation_externe(
  donnees,
  unite = NULL,
  age_cat = NULL,
  age_cat_mins = NULL,
  age_filtre_max = Inf,
  sexe = NULL,
  numerateur = NULL,
  denominateur = NULL,
  ref_externe_annee = NULL,
  ref_externe_code = "99",
  multiplicateur = 1
)
}
\arguments{
\item{donnees}{Un dataframe contenant les données à analyser. Les valeurs NA ne sont pas permises.}

\item{unite}{Un champ dans le dataframe qui indique l'unité d'analyse (par exemple, la région ou l'année).}

\item{age_cat}{Un champ dans le dataframe qui indique les groupes d'âge.}

\item{age_cat_mins}{Un vecteur contenant les âges minimaux de chaque groupe d'âge. Cet argument sert à grouper la variable d'âge des données externe de référence.}

\item{age_filtre_max}{Une valeur numérique indiquant le filtre d'âge maximum exclusif, Inf par défaut.}

\item{sexe}{Un champ dans le dataframe qui indique le sexe de chaque individu. Les valeurs de cette variable doivent être "M" et "F". Facultatif.}

\item{numerateur}{Un champ dans le dataframe qui représente le numérateur de la mesure de taux.}

\item{denominateur}{Un champ dans le dataframe qui représente le dénominateur de la mesure de taux.}

\item{ref_externe_annee}{Un champ dans le dataframe qui indique l'année de référence externe.}

\item{ref_externe_code}{Un champ dans le dataframe qui indique le code de référence externe.}

\item{multiplicateur}{Un nombre par lequel les taux standardisés sont multipliés (par exemple, 1000 pour obtenir des taux pour 1000 personnes).}
}
\value{
Une liste contenant deux éléments : "Resultat" (un dataframe contenant les résultats de la standardisation) 
        et "Details" (un dataframe contenant les données agrégées utilisées pour effectuer la standardisation).
}
\description{
Cette fonction applique une standardisation externe à des données agrégées, 
prenant comme référence un fichier de données pop_ref réprésentant la population du Québec 
stratifiée par âge et sexe, à de différentes années et niveaux géographiques. 
Les intervalles de confiance sont issues du package `PHEIndicatormethods`.
La standardisation externe doit être directe.
}
\examples{
\dontrun{
# Exemple d'utilisation de la fonction standardisation_externe

# Créer un dataframe exemple
donnees <- data.frame(
  expand.grid(region =c("Région 1", "Région 2"),
              groupe_age =c("20-29", "30-39", "40-49", "50-59", "60-69", "70-79"),
              sexe = c("F", "M")),
  numerateur = rpois(24, 10),
  denominateur = rpois(24, 200)
)

# Utiliser la fonction standardisation_externe
resultat <- standardisation_externe(
  donnees = donnees,
  unite = "region",
  age_cat = "groupe_age",
  age_cat_mins = c(20, 30, 40, 50, 60, 70),
  numerateur = "numerateur",
  denominateur = "denominateur",
  ref_externe_annee = 2022,
  ref_externe_code = "99",
  multiplicateur = 1000
)

# Afficher les résultats
print(resultat$Resultat)
}
}
\keyword{internal}
