% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisation_interne.R
\encoding{UTF-8}
\name{standardisation_interne}
\alias{standardisation_interne}
\title{Standardisation interne}
\usage{
standardisation_interne(
  donnees,
  unite = NULL,
  age_cat = NULL,
  sexe = NULL,
  autres_vars = NULL,
  numerateur = NULL,
  denominateur = NULL,
  reference_unite = NULL,
  methode = "directe",
  multiplicateur = 1
)
}
\arguments{
\item{donnees}{Un dataframe contenant les données. Les valeurs NA ne sont pas permises.}

\item{unite}{Une chaîne indiquant l'unité d'analyse.}

\item{age_cat}{Une chaîne indiquant la catégorie d'âge.}

\item{sexe}{Un champ dans le dataframe qui indique le sexe de chaque individu. Les valeurs de cette variable doivent être "M" et "F". Facultatif.}

\item{autres_vars}{Un vecteur optionnel d'autres variable(s) pour la stratification. Facultatif.}

\item{numerateur}{Un champ dans le dataframe qui représente le numérateur de la mesure de taux.}

\item{denominateur}{Un champ dans le dataframe qui représente le dénominateur de la mesure de taux.}

\item{reference_unite}{Une chaîne optionnelle indiquant l'unité de référence, NULL par défaut.}

\item{methode}{Une chaîne indiquant la méthode de standardisation ("directe" par défaut, ou "indirecte").}

\item{multiplicateur}{Un nombre par lequel les taux standardisés sont multipliés (par exemple, 1000 pour obtenir des taux pour 1000 personnes).}
}
\value{
Une liste avec deux éléments, "Resultat" et "Details". "Resultat" contient les résultats de 
        la standardisation, et "Details" contient le tableau de données agrégées.
}
\description{
Cette fonction applique une standardisation interne à des données agrégées. 
Dans le processus de standardisation, elle utilise soit la somme des données d'entrée comme population de référence
ou une unité spécifiée par `reference_unite`. La fonction ajuste pour des facteurs tels que le groupe d'âge, le sexe, et d'autres variables fournies
Les intervalles de confiance sont issues du package `PHEIndicatormethods`.
La standardisation interne peut-être directe ou indirecte.
}
\examples{
\dontrun{
# Exemple d'utilisation de la fonction standardisation_interne

# Créer un dataframe exemple
donnees <- data.frame(
  expand.grid(region =c("Région 1", "Région 2"),
              groupe_age =c("20-29", "30-39", "40-49", "50-59", "60-69", "70-79"),
              sexe = c("F", "M")),
  numerateur = rpois(24, 10),
  denominateur = rpois(24, 200)
)

# Utiliser la fonction standardisation_interne
resultat <- standardisation_interne(
  donnees = donnees,
  unite = "region",
  age_cat = "groupe_age",
  numerateur = "numerateur",
  denominateur = "denominateur",
  methode = "indirecte",
  multiplicateur = 1000
)

# Afficher les résultats
print(resultat$Resultat)
}

}
\keyword{internal}
